Relocate Dummy models
