+++ src/model/dummy.go
@@ -1,0 +1,27 @@
+// Copyright © 2020 Banzai Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package model
+
+// DummyClusterModel describes the dummy cluster model
+type DummyClusterModel struct {
+	ID                uint `gorm:"primary_key"`
+	KubernetesVersion string
+	NodeCount         int
+}
+
+// TableName sets the DummyClusterModel's table name
+func (DummyClusterModel) TableName() string {
+	return "dummy_clusters"
+}
